numerics builds

+++ geometry/barycentre_calculator_body.hpp
@@ -83,12 +83,10 @@ Point Barycentre(Point const (&points)[size]) {
       total += points[i] - origin;
     }
   }
-  // TODO(phl): The cast is necessary because `std::size_t` is not a field.
-  // It's somewhat annoying, though.  Can we fix this?
   if constexpr (additive_group<Point>) {
-    return total / static_cast<double>(size);
+    return total / size;
   } else {
-    return origin + total / static_cast<double>(size);
+    return origin + total / size;
   }
 }
